--- src/pages/checkout/delivery/delivery-type/mail-services/currier/index.js
@@ -1,4 +1,0 @@
-import { CurrierFirstStep } from './currier-first';
-import { CurrierSecondStep } from './currier-second';
-
-export { CurrierFirstStep, CurrierSecondStep };
